Add timezone, move userAgentData
