+++ packages/browser/src/plugins/segmentio/index.ts
@@ -12,8 +12,6 @@ import standard, { StandardDispatcherConfig } from './fetch-dispatcher'
 import { normalize } from './normalize'
 import { scheduleFlush } from './schedule-flush'
 import { SEGMENT_API_HOST } from '../../core/constants'
-import { clientHints } from '../../lib/client-hints'
-import { UADataValues } from '../../lib/client-hints/interfaces'
 
 type DeliveryStrategy =
   | {
@@ -52,11 +50,11 @@ function onAlias(analytics: Analytics, json: JSON): JSON {
   return json
 }
 
-export async function segmentio(
+export function segmentio(
   analytics: Analytics,
   settings?: SegmentioSettings,
   integrations?: LegacySettings['integrations']
-): Promise<Plugin> {
+): Plugin {
   // Attach `pagehide` before buffer is created so that inflight events are added
   // to the buffer before the buffer persists events in its own `pagehide` handler.
   window.addEventListener('pagehide', () => {
@@ -86,15 +84,6 @@ export async function segmentio(
       ? batch(apiHost, deliveryStrategy.config)
       : standard(deliveryStrategy?.config as StandardDispatcherConfig)
 
-  let userAgentData: UADataValues | undefined
-  try {
-    userAgentData = await clientHints(
-      analytics.options.highEntropyValuesClientHints
-    )
-  } catch {
-    userAgentData = undefined
-  }
-
   async function send(ctx: Context): Promise<Context> {
     if (isOffline()) {
       buffer.push(ctx)
@@ -106,10 +95,6 @@ export async function segmentio(
     inflightEvents.add(ctx)
 
     const path = ctx.event.type.charAt(0)
-
-    if (userAgentData && ctx.event.context) {
-      ctx.event.context.userAgentData = userAgentData
-    }
 
     let json = toFacade(ctx.event).json()
